removal of duplicate config move to beachbot's
